% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colours.R
\name{get_oxford_colour}
\alias{get_oxford_colour}
\alias{get_oxford_colours}
\title{Get named Oxford colours vector}
\usage{
get_oxford_colour(
  pattern = NULL,
  model = c("hex", "rgb", "cmyk", "pantone"),
  named = FALSE
)

get_oxford_colours(
  pattern = NULL,
  model = c("hex", "rgb", "cmyk", "pantone"),
  named = FALSE
)
}
\arguments{
\item{pattern}{Optional. A character value or vector to use as a search term.
Default is NULL in which case all the Oxford colours are returned.}

\item{model}{A character vector of colour model. Can be "rgb", "cmyk", "hex",
or "pantone". Default is "hex".}

\item{named}{Logical. Should the output be a named character value or
vector? Default is FALSE.}
}
\value{
A character value or vector of Oxford colour/s as per \code{model}
specification. If \code{named} is TRUE, returns a named character value or
vector.
}
\description{
Get named Oxford colours vector
}
\examples{
get_oxford_colours()
get_oxford_colours(model = "rgb")
get_oxford_colours(pattern = "lilac")
get_oxford_colours(pattern = c("lilac", "sage green"))

}
