% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nut_survey_map}
\alias{nut_survey_map}
\title{Example map data for showing map plotting with Oxford colours}
\format{
An sf object with 12 columns and 15 rows:\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{stateID} \tab State identifier \cr
   \emph{localityID} \tab Locality identifier \cr
   \emph{state_name} \tab State name \cr
   \emph{district} \tab District name \cr
   \emph{muac_gam} \tab Prevalence of global acute malnutrition by MUAC \cr
   \emph{muac_mam} \tab Prevalence of moderate acute malnutrition by MUAC \cr
   \emph{muac_sam} \tab Prevalence of severe acute malnutrition by MUAC \cr
   \emph{oedema} \tab Prevalence of severe acute malnutrition by oedema \cr
   \emph{dia} \tab Period prevalence of childhood diarrhoea \cr
   \emph{fev} \tab Period prevalence of childhood fever \cr
   \emph{bf} \tab Prevalence of continued breastfeeding \cr
   \emph{geom} \tab Geometry \cr
}
}
\usage{
nut_survey_map
}
\description{
Example map data for showing map plotting with Oxford colours
}
\examples{
nut_survey_map


}
\keyword{datasets}
