% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxford_palette.R
\name{oxford_brewer_palettes}
\alias{oxford_brewer_palettes}
\title{Oxford base map palettes}
\usage{
oxford_brewer_palettes()
}
\description{
Oxford base map palettes
}
\examples{
oxford_brewer_palettes()

}
