% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oxford_colours}
\alias{oxford_colours}
\title{Oxford colours based on University of Oxford's visual identity guidelines}
\format{
A data frame with 5 columns and 34 rows:

| **Variable** | **Description** |
| :--- | :--- |
| *name* | Official Oxford colour name |
| *rgb* | Three integers for the red, green, blue components of the RGB colour model |
| *cmyk* | Four integers for the cyan, magenta, yellow, and black components of the CMYK colour model |
| *hex* | Hexadecimal codes for corresponding colour |
| *pantone* | Pantone colour name |
}
\source{
https://communications.web.ox.ac.uk/communications-resources/visual-identity/identity-guidelines/colours
}
\usage{
oxford_colours
}
\description{
Oxford colours based on University of Oxford's visual identity guidelines
}
\examples{
oxford_colours

}
\keyword{datasets}
