% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxford_fonts.R
\docType{data}
\name{oxford_fonts}
\alias{oxford_fonts}
\alias{check_oxford_fonts}
\title{Oxford core and additional fonts}
\format{
An object of class \code{list} of length 7.
}
\usage{
oxford_fonts

check_oxford_fonts()
}
\value{
A list of available and not available fonts and a message indicating
availability status of Oxford fonts.
}
\description{
Oxford core and additional fonts

Check whether Oxford fonts (core and theme) are installed and available
}
\examples{
check_oxford_fonts()


}
\keyword{datasets}
