% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_palette.R
\name{oxford_palettes}
\alias{oxford_palettes}
\title{Complete list of Oxford colour palettes}
\usage{
oxford_palettes()
}
\description{
Complete list of Oxford colour palettes
}
\examples{
oxford_palettes()

}
