% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxford_palette.R
\name{oxford_theme_palettes}
\alias{oxford_theme_palettes}
\title{Oxford theme palettes from visual identity guidelines}
\usage{
oxford_theme_palettes()
}
\description{
Oxford theme palettes from visual identity guidelines
}
\examples{
oxford_theme_palettes()

}
