% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxfordtheme.R
\docType{package}
\name{oxfordtheme}
\alias{oxfordtheme}
\alias{oxfordtheme-package}
\title{Oxford Palette, Theme, and Theme Components}
\description{
Palette, theme, and theme components based on University of Oxford's visual
identity guidelines
\url{https://communications.web.ox.ac.uk/communications-resources/visual-identity/identity-guidelines}.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://oxford-ihtm.io/oxfordtheme/}
}

}
\author{
Ernest Guevarra (ernest@guevarra.io)
}
\keyword{internal}
