% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxthema.R
\docType{package}
\name{oxthema}
\alias{oxthema-package}
\alias{oxthema}
\title{Oxford Palette, Theme, and Theme Components}
\description{
Palette, theme, and theme components based on University of Oxford's visual
identity guidelines
\url{https://communications.web.ox.ac.uk/communications-resources/visual-identity/identity-guidelines}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://oxford-ihtm.io/oxthema/,https://github.com/OxfordIHTM/oxthema}
  \item Report bugs at \url{https://github.com/OxfordIHTM/oxthema/issues}
}

}
\author{
Ernest Guevarra

Greco Malijan
}
\keyword{internal}
