% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_palette.R
\name{create_palette_sequential}
\alias{create_palette_sequential}
\alias{create_palette_divergent}
\alias{create_palette_qualitative}
\alias{create_brewer_palette}
\title{Create new palettes based on Oxford palettes}
\usage{
create_palette_sequential(name, n)

create_palette_divergent(name, n)

create_palette_qualitative(name, n)

create_brewer_palette(
  name,
  n,
  type = c("sequential", "divergent", "qualitative")
)
}
\arguments{
\item{name}{Name of Oxford palette to use}

\item{n}{Number of colours desired/required. Oxford palettes have at least
5 colours. All colour schemes are derived from the University of Oxford
\href{https://communications.web.ox.ac.uk/communications-resources/visual-identity/identity-guidelines}{visual identity guidelines}.
#If NULL (default), use all colours.}

\item{type}{A character value for type of palette to use. Can be either
sequential, divergent, or qualitative.}
}
\value{
A character vector of desired/required colours with length equivalent
to \code{n}
}
\description{
These functions apply a similar approach used and demonstrated by
\href{https://colorbrewer2.org}{ColorBrewer} and has been patterned after the
syntax of the \href{https://cran.r-project.org/web/packages/RColorBrewer/index.html}{RColorBrewer}
package
}
\examples{
create_palette_sequential("blues", 5)
create_palette_divergent("brbg", 10)
create_palette_qualitative("dark", 5)
create_brewer_palette("blues", 5)

}
