% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_named_colours.R
\name{get_oxford_colour}
\alias{get_oxford_colour}
\alias{get_oxford_colours}
\title{Get named Oxford colours vector}
\usage{
get_oxford_colour(pattern = NULL, type = c("hex", "rgb", "cmyk", "pantone"))

get_oxford_colours(pattern = NULL, type = c("hex", "rgb", "cmyk", "pantone"))
}
\arguments{
\item{pattern}{Optional. A character value or vector to use as a search term.
Default is NULL in which case all the Oxford colours are returned.}

\item{type}{A character vector of colour type. Can be "rgb", "cmyk", "hex",
or "pantone". Default is "hex".}
}
\value{
A named vector of Oxford colour/s
}
\description{
Get named Oxford colours vector
}
\examples{
get_oxford_colours()
get_oxford_colours(type = "rgb")
get_oxford_colours(pattern = "lilac")
get_oxford_colours(pattern = c("lilac", "sage green"))

}
