% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxford_palettes.R
\name{oxford_theme_palettes}
\alias{oxford_theme_palettes}
\alias{oxford_brewer_palettes}
\title{Oxford theme palettes from visual identity guidelines}
\usage{
oxford_theme_palettes()

oxford_brewer_palettes()
}
\description{
Oxford theme palettes from visual identity guidelines

Oxford base map palettes
}
\examples{
oxford_theme_palettes()

oxford_brewer_palettes()

}
