% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ox_palettes.R
\docType{data}
\name{oxford_palettes}
\alias{oxford_palettes}
\title{Complete list of Oxford colour palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
oxford_palettes
}
\description{
Complete list of Oxford colour palettes
}
\keyword{datasets}
