% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxfordtheme.R
\docType{package}
\name{oxfordtheme}
\alias{oxfordtheme}
\alias{oxfordtheme-package}
\title{Oxford Theme and Theme Components}
\description{
Theme and theme components for 'ggplot2' based on University of Oxford's
visual identity guidelines
\url{https://www.ox.ac.uk/sites/files/oxford/media_wysiwyg/Oxford\%20Blue\%20LR.pdf}.
}
\author{
Ernest Guevarra (ernest@guevarra.io)
}
\keyword{internal}
