% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxfordtheme.R
\docType{package}
\name{oxfordtheme}
\alias{oxfordtheme}
\alias{oxfordtheme-package}
\title{Oxford Theme and Theme Components}
\description{
Theme and theme components for 'ggplot2' based on University of Oxford's
visual identity guidelines \url{https://communications.web.ox.ac.uk/communications-resources/visual-identity/identity-guidelines}
}
\author{
Ernest Guevarra (ernest@guevarra.io)
}
\keyword{internal}
