% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxford_theme_forest.R
\name{theme_oxford_forest}
\alias{theme_oxford_forest}
\alias{theme_heritage_forest}
\alias{theme_contemporary_forest}
\alias{theme_celebratory_forest}
\alias{theme_corporate_forest}
\alias{theme_innovative_forest}
\title{Oxford theme for \link{forestploter} package}
\usage{
theme_oxford_forest(
  base_size = 11.5,
  base_family = "Roboto",
  bg_col = get_oxford_colour("cool"),
  ci_col = get_oxford_colour("Oxford blue"),
  ci_fill = ci_col,
  refline_col = get_oxford_colour("red"),
  vertline_col = get_oxford_colour("ash"),
  summary_col = get_oxford_colour("ash"),
  summary_fill = summary_col,
  footnote_cex = 0.7,
  footnote_fontface = "plain",
  footnote_col = get_oxford_colour("royal"),
  title_col = get_oxford_colour("Oxford blue"),
  title_fontfamily = base_family,
  arrow_fill = get_oxford_colour("Oxford blue"),
  arrow_col = arrow_fill
)

theme_heritage_forest(
  base_size = 11,
  base_family = "Marcellus",
  bg_col = get_oxford_colour("sage"),
  ci_col = get_oxford_colour("Oxford blue"),
  ci_fill = ci_col,
  refline_col = get_oxford_colour("sienna"),
  vertline_col = get_oxford_colour("peach"),
  summary_col = get_oxford_colour("green"),
  summary_fill = summary_col,
  footnote_cex = 0.8,
  footnote_fontface = "plain",
  footnote_col = get_oxford_colour("green"),
  title_col = get_oxford_colour("Oxford blue"),
  title_fontfamily = base_family,
  arrow_fill = get_oxford_colour("Oxford blue"),
  arrow_col = arrow_fill
)

theme_contemporary_forest(
  base_size = 10.5,
  base_family = "Montserrat",
  bg_col = get_oxford_colour("cool"),
  ci_col = get_oxford_colour("Oxford blue"),
  ci_fill = ci_col,
  refline_col = get_oxford_colour("coral"),
  vertline_col = get_oxford_colour("potters"),
  summary_col = get_oxford_colour("aqua"),
  summary_fill = summary_col,
  footnote_cex = 0.7,
  footnote_fontface = "plain",
  footnote_col = get_oxford_colour("aqua"),
  title_col = get_oxford_colour("Oxford blue"),
  title_fontfamily = base_family,
  arrow_fill = get_oxford_colour("Oxford blue"),
  arrow_col = arrow_fill
)

theme_celebratory_forest(
  base_size = 10.5,
  base_family = "Noto Serif Display",
  bg_col = get_oxford_colour("lavender"),
  ci_col = get_oxford_colour("Oxford blue"),
  ci_fill = ci_col,
  refline_col = get_oxford_colour("plum"),
  vertline_col = get_oxford_colour("mauve"),
  summary_col = get_oxford_colour("viridian"),
  summary_fill = summary_col,
  footnote_cex = 0.7,
  footnote_fontface = "plain",
  footnote_col = get_oxford_colour("viridian"),
  title_col = get_oxford_colour("Oxford blue"),
  title_fontfamily = base_family,
  arrow_fill = get_oxford_colour("Oxford blue"),
  arrow_col = arrow_fill
)

theme_corporate_forest(
  base_size = 10.5,
  base_family = "Bebas Neue",
  bg_col = get_oxford_colour("sky"),
  ci_col = get_oxford_colour("Oxford blue"),
  ci_fill = ci_col,
  refline_col = get_oxford_colour("orange"),
  vertline_col = get_oxford_colour("sky"),
  summary_col = get_oxford_colour("royal"),
  summary_fill = summary_col,
  footnote_cex = 0.7,
  footnote_fontface = "plain",
  footnote_col = get_oxford_colour("royal"),
  title_col = get_oxford_colour("Oxford blue"),
  title_fontfamily = base_family,
  arrow_fill = get_oxford_colour("Oxford blue"),
  arrow_col = arrow_fill
)

theme_innovative_forest(
  base_size = 10.5,
  base_family = "Phudu",
  bg_col = get_oxford_colour("lavender"),
  ci_col = get_oxford_colour("Oxford blue"),
  ci_fill = ci_col,
  refline_col = get_oxford_colour("Oxford pink"),
  vertline_col = get_oxford_colour("vivid"),
  summary_col = get_oxford_colour("viridian"),
  summary_fill = summary_col,
  footnote_cex = 0.7,
  footnote_fontface = "plain",
  footnote_col = get_oxford_colour("viridian"),
  title_col = get_oxford_colour("Oxford blue"),
  title_fontfamily = base_family,
  arrow_fill = get_oxford_colour("Oxford blue"),
  arrow_col = arrow_fill
)
}
\arguments{
\item{base_size}{The size of text}

\item{base_family}{The font family to use for the text. Default to
main Oxford font \emph{Roboto}}

\item{bg_col}{Background colour for alternating rows. Default to
\emph{Oxford cool grey}.}

\item{ci_col}{Colour of the CI. A vector of colour should be provided for
the grouped forest plot. Default set to \emph{Oxford blue}}

\item{ci_fill}{Colour fill the point estimation. A vector of colour should be
provided for the grouped forest plot. Default to value of \code{ci_col}.}

\item{refline_col}{Line colour for the reference line. Default set to
\emph{Oxford red}.}

\item{vertline_col}{Line colour for the extra vertical line. Default set to
\emph{Oxford ash grey}.}

\item{summary_col}{Colour for borders of the summary diamond shape. Default
set to \emph{Oxford ash grey}}

\item{summary_fill}{Colour for filling the summary diamond shape. Default set
to value of \code{summary_col}.}

\item{footnote_cex}{Multiplier applied to font size for footnote.}

\item{footnote_fontface}{The font face for footnote.}

\item{footnote_col}{Colour of the footnote. Default set to
\emph{Oxford royal blue}.}

\item{title_col}{Colour of title. Default set to \emph{Oxford blue}}

\item{title_fontfamily}{Font family of title. Default set to value of
\code{base_family}.}

\item{arrow_fill}{Filling colour of the arrow head. Default set to
\emph{Oxford blue}.}

\item{arrow_col}{Line and text colour of the arrow. Default set to value of
\code{arrow_fill}.}
}
\value{
A list of specified theme parameters.
}
\description{
Oxford theme for \link{forestploter} package
}
\examples{
theme_oxford_forest()

}
